Creates Work component and moves Framework cards.
